De-duplicate some code between gateway and client
