+++ rust/connlib/libs/common/Cargo.toml
@@ -29,6 +29,7 @@ tracing-appender = "0.2"
 url = { version = "2.4.1", default-features = false }
 uuid = { version = "1.4", default-features = false, features = ["std", "v4", "serde"] }
 webrtc = { version = "0.8" }
+ring = "0.16"
 
 # Needed for Android logging until tracing is working
 log = "0.4"
